testStore/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
psycopg2==2.9.1